+++ hygen/_templates/source/new/source.ejs.t
@@ -1,0 +1,13 @@
+---
+to: src/data/sources/<%= h.changeCase.camelCase(name) -%>.js
+---
+const <%= h.changeCase.constantCase(name) %>_ID = '<%= h.changeCase.camelCase(name) %>';
+
+export default {
+  id: <%= h.changeCase.constantCase(name) %>_ID,
+  data: `
+    <%- data -%>
+
+  `,
+  type: '<%- type -%>',
+};