+++ template-sample-app/template/src/components/views/datasets/Datasets.js
@@ -14,7 +14,7 @@ import { setBottomSheetOpen, setError } from 'config/appSlice';
 
 const useStyles = makeStyles((theme) => ({
   title: {
-    padding: theme.spacing(3, 3, 1.5)
+    padding: theme.spacing(3, 3, 1.5),
   },
 
   content: {
@@ -61,6 +61,8 @@ function Datasets() {
       abortController.abort();
     };
   }, [credentials, dispatch]);
+
+  // Auto import useEffect
 
   const onParamsRefreshed = (oauthParams) => {
     if (oauthParams.error) {